refactor: separate topics into files
