+++ src/widgets/button/mod.rs
@@ -1,4 +1,6 @@
 mod button;
+mod event;
 mod theme;
+mod widget;
 
 pub use button::Button;
