+++ src/widgets/button/widget.rs
@@ -1,0 +1,91 @@
+use ratatui::buffer::Buffer;
+use ratatui::layout::Rect;
+use ratatui::style::Style;
+use ratatui::text::Line;
+use ratatui::widgets::Widget;
+
+use super::button::Button;
+
+impl<'a> Widget for Button<'a> {
+    fn render(self, area: Rect, buf: &mut Buffer) {
+        let (background, text, shadow, highlight) = self.colors();
+        buf.set_style(area, Style::new().bg(background).fg(text));
+
+        // render top line if there's enough space
+        if area.height > 2 {
+            buf.set_string(
+                area.x,
+                area.y,
+                "▔".repeat(area.width as usize),
+                Style::new().fg(highlight).bg(background),
+            );
+        }
+        // render bottom line if there's enough space
+        if area.height > 1 {
+            buf.set_string(
+                area.x,
+                area.y + area.height - 1,
+                "▁".repeat(area.width as usize),
+                Style::new().fg(shadow).bg(background),
+            );
+        }
+        // render label centered
+        let label = Line::from(self.label);
+        buf.set_line(
+            area.x + (area.width.saturating_sub(label.width() as u16)) / 2,
+            area.y + (area.height.saturating_sub(1)) / 2,
+            &label,
+            area.width,
+        );
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use ratatui::assert_buffer_eq;
+    use ratatui::style::Color;
+
+    use super::*;
+
+    const HIGHLIGHT_STYLE: Style = Style::new().fg(Color::White).bg(Color::Gray);
+    const TEXT_STYLE: Style = Style::new().fg(Color::Black).bg(Color::Gray);
+    const SHADOW_STYLE: Style = Style::new().fg(Color::DarkGray).bg(Color::Gray);
+
+    #[test]
+    fn render_default() {
+        let info = Button::new("Extract");
+        let mut buf = Buffer::empty(Rect::new(0, 0, 15, 3));
+
+        info.render(buf.area, &mut buf);
+
+        let mut expected = Buffer::with_lines(vec![
+            "▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔",
+            "    Extract    ",
+            "▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁",
+        ]);
+        expected.set_style(Rect::new(0, 0, 15, 1), HIGHLIGHT_STYLE);
+        expected.set_style(Rect::new(0, 1, 15, 1), TEXT_STYLE);
+        expected.set_style(Rect::new(0, 2, 15, 1), SHADOW_STYLE);
+
+        assert_buffer_eq!(buf, expected);
+    }
+
+    #[test]
+    fn render_active() {
+        let info = Button::new("Cancel").active(true);
+        let mut buf = Buffer::empty(Rect::new(0, 0, 15, 3));
+
+        info.render(buf.area, &mut buf);
+
+        let mut expected = Buffer::with_lines(vec![
+            "▔▔▔▔▔▔▔▔▔▔▔▔▔▔▔",
+            "    Cancel     ",
+            "▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁",
+        ]);
+        expected.set_style(Rect::new(0, 0, 15, 1), SHADOW_STYLE);
+        expected.set_style(Rect::new(0, 1, 15, 1), TEXT_STYLE);
+        expected.set_style(Rect::new(0, 2, 15, 1), HIGHLIGHT_STYLE);
+
+        assert_buffer_eq!(buf, expected);
+    }
+}
